pyproject.toml
[tool.poetry]
name = "opentelemetry-c-replayer"
version = "0.1.0"
description = "Read all opentelemetry-c CTF traces and send them to an OTel collector using OTLP for GRPC."
license = "MIT"
authors = [
  "Eya-Tom Augustin SANGAM <eya-tom-augustin.sangam@polymtl.ca>",
]
maintainers = [
  "Eya-Tom Augustin SANGAM <eya-tom-augustin.sangam@polymtl.ca>",
]
readme = "README.md"
homepage = "https://github.com/dorsal-lab/opentelemetry-c-replayer"
repository = "https://github.com/dorsal-lab/opentelemetry-c-replayer"
documentation = "https://github.com/dorsal-lab/opentelemetry-c-replayer"
keywords = ["LTTng", "OpenTelemetry", "Traces", "Spans"]
include = [
  "/src",
]

[tool.poetry.dependencies]
python = "^3.10"
opentelemetry-exporter-otlp-proto-grpc = "^1.15.0"
types-protobuf = "^4.21.0.2"
tqdm = "^4.64.1"
opentelemetry-semantic-conventions = "^0.39b0"

[tool.mypy]
namespace_packages = true
ignore_missing_imports = true
